chore: Move fingerprinting classes to a new package
