+++ core/src/main/java/io/naryo/application/configuration/revision/fingerprint/DefaultRevisionFingerprinter.java
@@ -1,4 +1,4 @@
-package io.naryo.application.common.revision;
+package io.naryo.application.configuration.revision.fingerprint;
 
 import java.util.Comparator;
 import java.util.List;
