+++ core/src/main/java/io/naryo/application/configuration/revision/fingerprint/RevisionFingerprinter.java
@@ -1,4 +1,4 @@
-package io.naryo.application.common.revision;
+package io.naryo.application.configuration.revision.fingerprint;
 
 import java.util.List;
 import java.util.UUID;
