+++ core/src/main/java/io/naryo/application/filter/revision/FilterRevisionFingerprinter.java
@@ -1,6 +1,6 @@
 package io.naryo.application.filter.revision;
 
-import io.naryo.application.common.revision.DefaultRevisionFingerprinter;
+import io.naryo.application.configuration.revision.fingerprint.DefaultRevisionFingerprinter;
 import io.naryo.domain.filter.Filter;
 import io.naryo.domain.filter.FilterNormalizer;
