+++ core/src/main/java/io/naryo/application/node/revision/NodeRevisionFingerprinter.java
@@ -1,6 +1,6 @@
 package io.naryo.application.node.revision;
 
-import io.naryo.application.common.revision.DefaultRevisionFingerprinter;
+import io.naryo.application.configuration.revision.fingerprint.DefaultRevisionFingerprinter;
 import io.naryo.domain.node.Node;
 import io.naryo.domain.node.NodeNormalizer;
